refactor(tasks): separate tasks into different namespaces
