+++ tasks/common.py
@@ -1,0 +1,1 @@
+PIPENV_PREFIX = "pipenv run"
